PLT-525 - Delegated Access - review fix (#42)
* PLT-525 - Delegated Access - review fix

* PLT-525 - Delegated Access - review fix

* PLT-525 - Delegated Access - review fix

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppspersononprobationuserapi/jpa/entity/DelegatedAccessPermissionEntity.kt
@@ -16,10 +16,10 @@ data class DelegatedAccessPermissionEntity(
   val id: Long?,
 
   @Column(name = "delegated_access_id")
-  var delegatedAccessId: Int,
+  var delegatedAccessId: Long,
 
   @Column(name = "permission_id")
-  var permissionId: Int,
+  var permissionId: Long,
 
   @Column(name = "when_granted")
   val granted: LocalDateTime,
